[🚚rename]: Participants를 백엔드와 동일하게 Members로 변경
