+++ src/components/MemberCard/index.ts
@@ -1,0 +1,2 @@
+export { default as MemberCard } from './MemberCard';
+export { default as StatusTag } from './StatusTag';
