--- src/components/ParticipantsCard/index.ts
@@ -1,2 +1,0 @@
-export { default as ParticipantsCard } from './ParticipantsCard';
-export { default as StatusTag } from './StatusTag';
