+++ src/pages/RoomPage/RoomViews/index.ts
@@ -1,3 +1,3 @@
-export { default as Participants } from './Participants/Participants';
+export { default as Members } from './Members/Members';
 export { default as RoomHeader } from './RoomHeader/RoomHeader';
 export { default as TestInfo } from './TestInfo/TestInfo';
